% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_dataset_features.R
\name{add_dataset_features}
\alias{add_dataset_features}
\title{Add extra features to dataset. This is for feature expansion and modelling purposes These values will be highly correlated with
each other so these need to be filtered when modelled}
\usage{
add_dataset_features(dat, tgt_ori_col_name, derived_codes)
}
\arguments{
\item{dat}{data table of features}

\item{tgt_ori_col_name}{column name that will be used as the target}

\item{derived_codes}{the derived codes. this is used to add extra data for the dervied forecast. Omitted if not present}
}
\value{
data.table
}
\description{
Add extra features to dataset. This is for feature expansion and modelling purposes These values will be highly correlated with
each other so these need to be filtered when modelled
}
\examples{
 st_ft <- SteelForecast::add_dataset_features(st, tgt_ori_col_name = "MB_STE_0028", derived_codes = "MB_STE_0892")
}
