% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corr_cutoff.R
\name{corr_cutoff}
\alias{corr_cutoff}
\title{Wrapper for caret findCorrelation. Cuts off features with A numeric value for the
pair-wise absolute correlation cutoff above a certain amount.}
\usage{
corr_cutoff(X, cutoff = 0.9, verbose = FALSE)
}
\arguments{
\item{X}{the feature set to run cutoff on}

\item{cutoff}{cutoff value}

\item{verbose}{printing details}
}
\value{
matrix
}
\description{
Wrapper for caret findCorrelation. Cuts off features with A numeric value for the
pair-wise absolute correlation cutoff above a certain amount.
}
\examples{
corr_cutoff(X = matrix)
}
