% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_step_selection.R
\name{create_step_selection}
\alias{create_step_selection}
\title{Forward selection process}
\usage{
create_step_selection(
  train_X_all,
  test_X_all,
  train_y,
  validation,
  cutoff_val = 0.9,
  nvmax = 20,
  features = nvmax,
  method = "forward"
)
}
\arguments{
\item{train_X_all}{the training set}

\item{test_X_all}{the test set only gets amended if validation is true}

\item{train_y}{the target}

\item{validation}{are we testing the accuracy of the forecast or just running the forecast and getting the
best features?}

\item{nvmax}{max features to consider for the forward selection}

\item{features}{the number of features to}

\item{method}{step method}

\item{cutoff}{a correlation feature cutoff is done before the forward selection to prevent
singular matricies. This is how severe to be with the correlated valeus}
}
\value{
data.table
}
\description{
Forward selection process
}
\examples{
SteelForecast::create_step_selection(train_X, test_X, y, validation = TRUE)
}
