% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_synthetic_data.R
\name{create_synthetic_data}
\alias{create_synthetic_data}
\title{Creates synthetic data for any data table that has na values contained}
\usage{
create_synthetic_data(
  dat,
  plot = F,
  verbose = T,
  exclude_cols = NULL,
  dte_col,
  cutoff_val = 0.8,
  nvmax = 10,
  y_only = FALSE
)
}
\arguments{
\item{dat}{data table of data}

\item{plot}{whether to add the plot data to the return ouput}

\item{verbose}{printing values}

\item{exclude_cols}{which columns to exclude}

\item{dte_col}{the column for the date values in the timeseries}

\item{cutoff_val}{cutoff vlaue for the forward selection}

\item{nvmax}{the max of the forward selection}

\item{y_only}{if true then only usees the y value for the forecast with no regressors. Use this
for the sake of speed}
}
\value{
data.table
}
\description{
Works by forecasting backwards fort he nas using regressers if possible
method of dealing with nas
features selected with forward selection
Uses and unsupervised method of forecasting so doesn't check how accurate it is
It is only to guess some values for the variation and so the feature could be used in the forecast
The testing of the feature importance happens in the main model selection
}
\examples{
ceic_synth_list <- create_synthetic_data(ceic, plot = T, verbose = verbose, dte_col = "Mth")
ceic_synth <- ceic_synth_list[[2]]
ceic_synth_plot <- ceic_synth_list[[1]]
}
