% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_construction_eurostat.R
\name{get_construction_eurostat}
\alias{get_construction_eurostat}
\title{Pulls construction data from euro stat}
\usage{
get_construction_eurostat()
}
\value{
data.table
}
\description{
Pulls construction data from euro stat
}
\examples{
get_construction_eurostat()
}
