% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sentiment_eurostat.R
\name{get_sentiment_eurostat}
\alias{get_sentiment_eurostat}
\title{Gets sentiment data from Eurostat}
\usage{
get_sentiment_eurostat()
}
\value{
dasta.table
}
\description{
Gets sentiment data from Eurostat
}
\examples{
see get_steel_data
}
