% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_world_bank.R
\name{get_world_bank}
\alias{get_world_bank}
\title{Gets the worldbank data from the website}
\usage{
get_world_bank(
  url1 =
    "https://thedocs.worldbank.org/en/doc/5d903e848db1d1b83e0ec8f744e55570-0350012021/related/CMO-Historical-Data-Monthly.xlsx",
  forecast_date
)
}
\arguments{
\item{url1}{url to world bank data}
}
\value{
data.table
}
\description{
Gets the worldbank data from the website
}
\examples{
See get_steel_data
}
